source: outline `code_gen_helpers` from `model`
