+++ source/model.py
@@ -112,6 +112,11 @@ from .tools import (
 from collections import (
     deque
 )
+from .code_gen_helpers import (
+    gen_init_string,
+    gen_function_declaration_string,
+    gen_array_declaration,
+)
 
 
 # List of coding style specific code generation settings.
@@ -2784,50 +2789,6 @@ class EnumerationElementDeclaration(SourceChunk):
     separ = separ
             )
         )
-
-
-def gen_array_declaration(array_size):
-    if array_size is not None:
-        if array_size == 0:
-            array_decl = "[]"
-        elif array_size > 0:
-            array_decl = '[' + str(array_size) + ']'
-    else:
-        array_decl = ""
-    return array_decl
-
-
-def gen_function_declaration_string(indent, function,
-    pointer_name = None,
-    array_size = None
-):
-    if function.args is None:
-        args = "void"
-    else:
-        args = ",@s".join(a.declaration_string for a in function.args)
-
-    return "{indent}{static}{inline}{ret_type}{name}(@a{args}@c)".format(
-        indent = indent,
-        static = "static@b" if function.static else "",
-        inline = "inline@b" if function.inline else "",
-        ret_type = function.ret_type.declaration_string,
-        name = function.c_name if pointer_name is None else (
-            "(*" + pointer_name + gen_array_declaration(array_size) + ')'
-        ),
-        args = args
-    )
-
-
-def gen_init_string(type, initializer, indent):
-    init_code = ""
-    if initializer is not None:
-        raw_code = type.gen_usage_string(initializer)
-        # add indent to initializer code
-        init_code_lines = raw_code.split('\n')
-        init_code = "@b=@b" + init_code_lines[0]
-        for line in init_code_lines[1:]:
-            init_code += "\n" + indent + line
-    return init_code
 
 
 def gen_function_decl_ref_chunks(function, generator):
